AllActivityPage 디자인 변경
