+++ src/styles/List.jsx
@@ -1,9 +1,12 @@
 import styled from '@emotion/styled';
 
 const List = styled.ul({
+  display: 'grid',
+  gridTemplateColumns: 'repeat(3, 1fr)',
   listStyle: 'none',
   margin: '0',
   padding: '0',
+  paddingBottom: '50px',
 });
 
 export default List;
