+++ src/styles/Title.jsx
@@ -1,11 +1,9 @@
 import styled from '@emotion/styled';
 
 const Title = styled.h2({
-  paddingLeft: '20px',
-  paddingBottom: '10px',
-  color: '#FFB61D',
-  fontSize: '30px',
-  borderBottom: '2px solid #FFB61D',
+  margin: '5% 0',
+  textAlign: 'center',
+  fontSize: '40px',
 });
 
 export default Title;
